search/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.10
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
et-xmlfile==1.1.0
joblib==1.3.2
kombu==5.3.5
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
